rewritten, but the same..
i hadd some issues with the date not updating as it should. so i rewrote some of the code, made it more easier to use, and reliable.

i think that, how it is now, is more easy to catch up, whats doing what, and where to make changes if you want.